+++ src/api-hooks/MyPage.ts
@@ -1,9 +1,24 @@
 import { httpClient } from '.';
 import { useQuery } from '@tanstack/react-query';
-import { IMyArticleRequest, IMyArticleResponse, IMyNoticeResponse } from '../@types/MyPage';
+import {
+  IMyArticleData,
+  IMyArticleRequest,
+  IMyArticleResponse,
+  IMyNoticeData,
+  IMyNoticeResponse,
+} from '../@types/MyPage';
+import { useEffect, useState } from 'react';
+import toast from 'react-hot-toast';
 
-export function useGetMyInfoRequest(params: IMyArticleRequest, isEnabled?: boolean) {
-  return useQuery(
+export function useGetMyArticleRequest(params: IMyArticleRequest, isEnabled?: boolean) {
+  const [myArticleList, setMyArticleList] = useState<IMyArticleData[] | null>([]);
+  const [totalPosts, setTotalPosts] = useState(1);
+
+  const {
+    data,
+    error,
+    refetch: myArticleRefetch,
+  } = useQuery(
     [`/admin/article/my?page=${params.page}`],
     () =>
       httpClient<IMyArticleResponse>({
@@ -12,10 +27,28 @@ export function useGetMyInfoRequest(params: IMyArticleRequest, isEnabled?: boole
       }),
     { enabled: isEnabled },
   );
+
+  useEffect(() => {
+    if (data) {
+      setMyArticleList(data.data.articleList);
+      setTotalPosts(data.data.totalPage);
+    } else if (error) {
+      toast.error('내 글 목록을 불러오는데 실패했습니다..');
+    }
+  }, [data, error]);
+
+  return { myArticleList, totalPosts, myArticleRefetch };
 }
 
 export function useGetMyNoticeRequest(params: IMyArticleRequest, isEnabled?: boolean) {
-  return useQuery(
+  const [myNoticeList, setMyNoticeList] = useState<IMyNoticeData[] | null>([]);
+  const [totalPosts, setTotalPosts] = useState(1);
+
+  const {
+    data,
+    error,
+    refetch: myNoticeRefetch,
+  } = useQuery(
     [`/admin/notice/my?page=${params.page}`],
     () =>
       httpClient<IMyNoticeResponse>({
@@ -24,4 +57,15 @@ export function useGetMyNoticeRequest(params: IMyArticleRequest, isEnabled?: boo
       }),
     { enabled: isEnabled },
   );
+
+  useEffect(() => {
+    if (data) {
+      setMyNoticeList(data.data.noticeList);
+      setTotalPosts(data.data.totalPage);
+    } else if (error) {
+      toast.error('내 공지사항 목록을 불러오는데 실패했습니다..');
+    }
+  }, [data, error]);
+
+  return { myNoticeList, totalPosts, myNoticeRefetch };
 }
