+++ src/components/features/ArticleModifyModal/index.tsx
@@ -12,14 +12,13 @@ import { TitleInputSection, ButtonContainer } from './style';
 import { useUpdateInfoRequest, useGetInfoRequest } from '../../../api-hooks/Info';
 import { useUpdateNoticeRequest, useGetNoticeRequest } from '../../../api-hooks/Notice';
 import { useGetAllCategoryRequest } from '../../../api-hooks/InfoCategory';
-import { useGetMyInfoRequest, useGetMyNoticeRequest } from '../../../api-hooks/MyPage';
+import { useGetMyArticleRequest, useGetMyNoticeRequest } from '../../../api-hooks/MyPage';
 import { useUploadImageRequest } from '../../../api-hooks/ImageUpload';
 import toast from 'react-hot-toast';
 import { useSelectedArticleAtom } from '../../../store/articleAtom';
 import { useCurrentLocationAtom } from '../../../store/currentLocationAtom';
 import { Flex } from '@chakra-ui/react';
 import { ThumbnailUploadUI } from '../../ui/ThumbnailUploadUI';
-import { ICategoryData } from '../../../@types/InfoCategory';
 
 type EditorInstance = Editor | null;
 
@@ -67,14 +66,14 @@ function ArticleModifyModal({ location, isModalOpen, setIsModalOpen, currentPage
     false,
   );
 
-  const { refetch: myArticleRefetch } = useGetMyInfoRequest(
+  const { myArticleRefetch } = useGetMyArticleRequest(
     {
       page: currentPage - 1,
     },
     false,
   );
 
-  const { refetch: myNoticeRefetch } = useGetMyNoticeRequest(
+  const { myNoticeRefetch } = useGetMyNoticeRequest(
     {
       page: currentPage - 1,
     },
