+++ src/pages/MyPage/MyArticlePage/index.tsx
@@ -1,63 +1,15 @@
 import { Tabs, TabList, TabPanels, Tab, TabPanel } from '@chakra-ui/react';
 import RightContainer from '../../../components/layout/RightContainer';
-import Info from './Info';
-import Notice from './Notice';
-import { useSelectedArticleAtom } from '../../../store/articleAtom';
-import { useEffect } from 'react';
-import { useCurrentLocationAtom } from '../../../store/currentLocationAtom';
+import useMyArticle from './useMyArticle';
 
 function MyArticlePage() {
-  const { resetAtom } = useSelectedArticleAtom();
-  const { setCurrentLocation } = useCurrentLocationAtom();
-
-  useEffect(() => {
-    setCurrentLocation('작성한 정보');
-  }, []);
-
-  useEffect(() => {
-    resetAtom();
-  }, []);
-
-  const handleCurrentLocation = (location: string) => {
-    setCurrentLocation(location);
-    resetAtom();
-  };
-
-  const ACCOUNT_MENU_DEFAULT = [
-    {
-      id: 1,
-      tabName: '작성한 정보',
-    },
-  ];
-
-  const ACCOUNT_MENU_ADMIN = [
-    {
-      id: 1,
-      tabName: '작성한 정보',
-    },
-    {
-      id: 2,
-      tabName: '작성한 공지사항',
-    },
-  ];
-
-  const ACCOUNT_SCREEN_DEFAULT = [
-    {
-      id: 1,
-      screen: <Info />,
-    },
-  ];
-
-  const ACCOUNT_SCREEN_ADMIN = [
-    {
-      id: 1,
-      screen: <Info />,
-    },
-    {
-      id: 2,
-      screen: <Notice />,
-    },
-  ];
+  const {
+    ACCOUNT_MENU_DEFAULT,
+    ACCOUNT_MENU_ADMIN,
+    ACCOUNT_SCREEN_DEFAULT,
+    ACCOUNT_SCREEN_ADMIN,
+    handleCurrentLocation,
+  } = useMyArticle();
 
   return (
     <RightContainer title={'내가 작성한 글 관리'}>
